Add timestamps to screams
I'd like to know when screams were posted. That seems important.
Unfortunately, I was unable to find a nice way to make this work using
hidden fields or anything, so I needed to use the same technique used in
[Ask] and create a new datatype to represent the form fields. Then I'm
able to supply a default creation time to the scream when I parse it out
of those fields.

I'm also setting the timestamp as the primary sort key, falling back to
the id in the case of a tie (unlikely).