Require authentication for screams
We want to make sure Joe Schmo isn't able to scream on my behalf. I've
decided to use yesod-auth-hashdb for the authentication, because all I
actually care about is one single user (me) and I'm happy to manually
create my credentials in the database.

I'm exposing `pwstore-fast` in order to be able to import it in ghci so
that I can generate the correct password hash.