Don't let users post screams longer than 280 chars
We want to make sure we're forcing users (me) to be brief. To this end,
we should keep track of the character count and explicitly disallow
screams longer than 280 chars. This length is arbitrary, but it's
consistent with what micro.blog will be expecting, so it seems
reasonable.

Additionally, we can keep track of if the scream is also a valid tweet.
This is useful info to have since micro.blog will (eventually) be
handling crossposting to twitter.