Get character count based on rendered markdown
Using the plain text of our scream to determine character count didn't
actually make sense. Microblog posts can have simple html embedded in
them, and we'd like to be able to take advantage of this. For example,
the scream `[foo](http://example.com)` would have had a character count
of 25 under the previous counting method, but it should _actually_ have
a character count of 3, since the only text displayed is `foo`.

To solve this, we (unfortunately) need to turn to JavaScript. We can
pull in a javascript markdown lib, and perform on-the-fly rendering of
our input text into an off-screen div. Then, we can just get the
innerText of this div and use _that_ to count the number of characters
in our scream.

I might look to minimize the javascript lib in the future, and there's a
_slight_ risk that the js lib renders things differently than
yesod-markdown (and vice versa), but we can cross those bridges when we
come to them.