Display a simple scream detail
This is pretty basic right now, but we just want to simply be able to
display a specific scream at /#ScreamId. This will give us nice little
URLs like `micro.gordonfontenot.com/563` and whatnot.

This route annoyingly conflicts with _literally every other route_ so we
need to use `!` to show that we're aware of that.