Refactor NewScream to use addScript
Instead of dumping these on the page directly, it makes more sense for
us to use the `addScript` features already exposed by Yesod. This lets
us use a nice typesafe route, while also centralizing the logic for
adding these dependencies to the page.

Additionally, since we're now doing extra work and we want to stay
consistent, we're extracting out a new little helper function for
rendering the new scream page with a given form and enctype.