Add some initial styles for the root view
We're going to be taking the base of our style from the main website, so
link those into the default layout wrapper. We then want to layer on
microblog-specific styles to refine the look and feel.

I'm also updating the overall layout to mimic the main website's layout
a bit more.