Refactor index and detail to use common markup
We want our screams to be displayed consistently between the detail and
the index. To make this happen, we can export a new function from
ScreamDetail that represents a widget for a single scream. This function
can then be imported into Home, where it can be reused in the (now
greatly simplified) hamlet file. Now, any changes to the detail will be
reflected in the index and that's one less thing I need to worry about.