Paginate screams
We don't want to display all 18k screams on the root once I
import those into the database. That would be horrific. Instead, we
should paginate the root so that it only shows 30 screams at a time.

This is made _dead simple_ using yesod-paginator from the imitable Pat
Brisbin. My only issue with the lib is that it does a little _too_ much
for me. I just want to have previous and next links, I don't need the
ability to jump to an individual page.

So, to accomplish this, I'm going to wrap yesod-paginator in my own
module that exposes a simpler version of the paginator widget, only
containing links for Older and Newer screams when appropriate. _Most_ of
this is taken directly from yesod-paginator so it should be trivial to
contribute back upstream if it's wanted.