Add simple imageURL prop to Screams
This is just populated with the filename for the image attached to a
scream for right now. The next step will be to flesh out the S3 upload
process, at which point we will have a usable URL for the images.
However, this _does_ get the types in place so that the project builds
and the right types will be persisted in the right place.