Various cleanups
I needed to mess with the app settings in a different branch and I felt
completely lost. As an effort to understand more about what was going on
in here, I decided to go ahead and go through these files and clean them
up a bit. This was mainly just removing comments and unused settings and
reformatting imports.

I took a bunch of cues from https://github.com/pbrisbin/tee-io when
figuring out what to keep/remove.