Upload image attachments to S3
This is a doozy of a commit, but this fleshes out our S3 helper to
upload images to S3 when we create a scream. I think the identifier
stuff is probably a bit naive right now, and I should probably flesh
that out before I move this into production. But this works as you'd
expect! Images are uploaded to S3 (credentials are specified by the ENV,
and the bucket name is stored in the settings) and the url for the image
is added to the Scream.