Use Stack lts 8.0
This lts was just released, and includes ghc 8.0.2 as the compiler. This
means the compilation bugs we were working around with the ghc/resolver
configurations are now fixed, and so we can just reference the lts
directly. Happy days!

Note that because `yesod-markdown` has an upper bound of 1.18 for
pandoc, we need to define an explicit version of pandoc as well as a
number of pandoc dependencies in order to keep using it. I've submitted
a PR upstream to bump that upper bound.

Also note that amazonka has been upgraded, so we no longer need to
explicitly specify the region. Instead, it will be discovered by the
server-side config.