Add deployment process using docker
This adds a series of scripts and configs to let us deploy our app to
heroku using docker. It uses the docker registry stuff that Heroku
exposes to let us push pre-built containers directly to heroku itself.
It also uses a two phase build process the reduce the container size.
The first phase builds the app itself, and so that container includes
stack and ghc and all of the other build tools needed to compile the
app. The second phase _doesn't_ need all of that crap, and so simply
copies in the built app and adds any runtime dependencies needed to run
the server.

The result is an _incredibly_ fast build/deployment process (after the
initial build) and dead-simple deployment.