Embed css into the app
Unfortunately, since my main website isn't hosted with ssl, I can't link
to the css from a website that _is_ hosted with ssl. This means that
on production, our app isn't able to see the css and so looks like shit.
Duplicating the css seems... not ideal, but it's better than changing
the entire hosting structure for my main website just to handle this one
case.