Combine image fetching functions
Instead of duplicating this logic, we can simply reuse the existing
logic for fetching images for multiple screams, and pass the single
scream as a single element list. This has the same behavior, but if the
logic around this ever changes, we'll only have to update it in one
place.