Don't set `-Werror` on the `task` executable
I'm using `undefined` a lot to work through stuff and it drives me crazy
to have it fail to compile, instead of just showing me the warnings. I
might re-enable this some time in the future, but for now I want to
disable it for the sake of development speed.