Add the ability to import a tweet into the db
This fleshes out the `task import` command so that it actually imports a
tweet into the database.

This command expects a pre-formatted JSON file with a specific name, and
images in a specific folder. Specifically, the expected format is:

```
parent-directory
├── entry.json
└── photos
    ├── 0.jpg
    └── 1.jpg
```

The entry.json file should be a single element json file with the
following structure:

```json
{
  "id": "804354572831641600",
  "text": "ffs",
  "created_at": "2016-12-01T16:00:49Z"
}
```

Any other fields will be ignored. The `id` should be the tweet id, the
`text` should be the raw body of the tweet, and `created_at` is the
published date.

Any photos found in the `photos` directory will be uploaded to S3 and
attached to the scream created from the imported tweet.