Only import the photos if the directory exists
We won't _always_ have a photos directory, and if we don't, listing the
contents will result in an error. To fix this we can first check to make
sure that the directory exists, and _then_ list the contents if so.

I'm like 99% sure there's an easier way to do this.