Truncate URLs over 30 characters
Sometimes I like to post a bare link, instead of using markdown. When I
do, it can get ugly, since bare urls cane be long and unruly and run off
the edge of the page.

To fix this, I'm going to go ahead and truncate the display text for any
bare url to 30 characters, and indicate that it has been truncated by
appending an ellipses. This will improve the display immensely, IMO.