Add link to tweet
If a scream has been cross-posted to Twitter, it would be nice to link
to the tweet so that people can interact with it there if they would
like. We can do this same thing for other syndication destinations in
the future (Facebook, for example). For now we'll just worry about
linking to the tweet, since that's the info we currently have.