Step back and use plain text for twitter links
I liked this idea, but ultimately it feels really invasive. I also worry
about how this would look with a gaggle of icons for any arbitrary
number of social networks hanging out on every post.

Instead, lets stick with our plain text aesthetic and use a simple `tw`
link at the end of the timestamp. This will be easy to expand to `fb`
and maybe `mb` if needed, and does a better job of blending in with the
existing design.