Use RepRss for our content type
If we use `Html` as our content type, the mime type ends up incorrect
when we serve this route. We'd like it to use an rss mime type instead.

The nicest way to accomplish this is probably to reuse the RepRss
content type defined by yesod-newsfeed. This type uses the correct
mimetype and has the required typeclass instances. We could duplicate
this info ourselves, but since yesod-newsfeed is already a transitive
dependency, we might as well just use it directly.