Save twitter credentials in the ENV
Since we don't need to authenticate with multiple twitter accounts, we
can skip a bunch of the hard parts around OAuth and just store our
credentials in the ENV.

To make this work with our app, it means we should add this config to
the app settings. We'll go ahead and define a custom type to hold onto
this stuff.