Add helper functions for constructing auth types
To be able to sign our requests properly, we need to make sure we have
an OAuth type as well as a Credential type constructed. These are a bit
involved, so it probably makes the most sense to extract this stuff into
helper functions that we can use once we actually sign the requests.