Add base function for creating requests
We need to make sure our requests for the twitter API are formatted
properly. To help us do this, we will centralize the creation of these
requests. This central function will use an Endpoint type to determine
where it should hit, what the payload should be, and how it should make
the request. But all requests will be signed, and will be forced to use
ssl.