Upload images to Twitter
If we have any images for our scream, we want to upload them to Twitter
and then collect the result. To do this, we'll create a new endpoint for
the uploads that accepts the raw image data.

We'll also create a new type to represent the result of the request
because annoyingly Twitter decided to use inconsistent JSON keys between
their different endpoints.