Simplify request creation logic
Previously, we were using `SimpleQuery` inside our Endpoint type, but
ultimately that meant pushing the work of creating those into the
endpoint creation. Ultimately, I think that's probably too noisy, and
that we can get by with a simple `(Text, Text)` representation of these
queries.

This leaves us in a bind though, because our upload endpoint is going to
deal with ByteString already. Moving that back and forth from Text to
ByteString seems like a waste of time, so I'd like to try to avoid that.

We can solve this issue by conflating the request method and the body
data into a sum type. We can then use this sum type to not only
differentiate between the different body formats, but also to determine
how to encode the particular body types for the request.