Consolidate feeds into single endpoint
We can actually use Yesod to serve up multiple representations of the
same content, based on the requested content type (via the Accept
header). This is a neat feature that we should definitely use.

To do this, we're going to restructure our feed construction into a more
data-centric approach, which I actually like better anyway. We'll have
some new data types and then we can implement `ToJSON` instances for
those types, which will automatically get us to the final JSON response.

The XML code actually remains largely unchanged, except that we're now
dealing with this new Feed type instead of the FeedItem type we were
using before.

We'll also take this opportunity to make dealing with these populated
scream types a bit nicer by defining a type alias for them, and then
adding some helper functions for dealing with them directly. This is
essentially what our old FeedItem type was in previous commits, but this
is a bit lighter weight.