Ensure that we're using FeedItems inside template
During refactoring, we accidentally left this template as referencing
'items' instead of pointing it at the feed's items. This meant that we
didn't have access to stuff we had already calculated, like the html
content and the timestamp.

To fix this, we can update our template to refer to the feed items
coming from the feed, and then we can ensure that this reference isn't
even available by one-lining our feed generation function.