Use Twitter's algorithm for character counts
Twitter does some special things like URL shortening with a constant
character count (23). This means that just counting the number of
characters in our rendered markdown could have _wildly_ different
character counts. We could either be showing a count that is way over if
we submit a super long URL, or way under if we have a short url.

To fix this, we can use the handy twitter-text javascript library, which
Twitter provides to do this exact kind of thing:

https://github.com/twitter/twitter-text/tree/master/js

With this lib in place, we just need to call the new specialized count
function and we'll end up with an accurate character count.