Use Pandoc for generating plain text screams
Previously, we were using hxt for generating the plain text versions of
our screams. This meant going from Markdown (Text) to Pandoc to Html to
Text. On top of being fairly inefficient, it also meant that during the
round trip we lost any newline characters that we had added to the text.

To simplify (and improve) things, we can instead use Pandoc directly to
print to plain text. This is possible out of the box by using
`writePlain` and a very simple set of writer options. However, this does
have one unexpected side effect of not stripping inline styles (even
though the docs say that it does). This means that `_Hello_ __world__`
would render as `_Hello_ WORLD`. I'd rather it just strip these out
entirely, so I'm adding a simple filter into the mix that strips these
elements from the document before writing.

I'm also taking the opportunity to rename this to `plainText` which I
like a bit more.