Remove yesod-markdown dependency
We were already duplicating a good chunk of the yesod-markdown lib
because we wanted to customize our reading options. As it turns out,
with a bit more code we can remove the dependency on the lib entirely,
which will have the added benefit of removing a bit more implicit
behavior, and expose our preferences directly.

While we're in here, I'm going to go ahead and reorganize things and
explicitly declare our imports/exports. This will make refactoring
easier in the future.