Don't escape encoded markup
Yesod defaults to escaping markup when using interpolation in Hamlet. This is
a good default, but we've actually already converted our html content to
rendered html, so escaping it here is counter productive.

To solve this, we can convert from `Text` to `Html` explicitly ourselves by
specifying that the content is already pre-escaped.

This will fix the html rendering in our RSS feeds.